Fixed crash, added Remington ACR support, added custom weapon tutorial
wpn_stats.lua and wpn_parts.lua now both contain a small "tutorial" on how to add support for custom weapons to the mod. Please submit a PR if you add one, I'd love to add it!

This update also fixes a crash in CopBrain when a new person joins.